Start and end connections based on "inCall" flags changes
Until now the connections were started when a participant joined with
audio and/or video (or, if there was no HPB, if the self participant had
audio and/or video), and ended when a participant left the call. If a
participant starts or stops publishing while being in the call, either
because the user now selected a microphone or camera or because the
participant was granted permissions, the participant force reconnected
to the call (leave and join it again), which ensured that all other
participants started or ended the connections as needed. To avoid having
to force reconnect and have a smoother call experience now the clients
are expected to keep track of the "inCall" flags and start and end the
connections as needed with the existing session.

Note that if the client is used against an older Talk version everything
would still work as before; as the other participant leaves and joins
the call again with a different session when needed the call flags would
not be updated for the previous session (other than to leave it), so the
new code to handle the updated call flags should have no effect (and,
even if it does, the connection would be established as needed against
the new session once the participant joins again).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>