improve voice message quality
first it was tried with sampling rate 44100 but this showed an error on safari when try to playback.

by direct comparison to 44100, 22050 has lower quality but it's totally okay compared to the quality before when no sampling rate was set. I'm not sure why it failed on safari, because the iOS app also uses 44100.

see https://github.com/nextcloud/talk-android/pull/2714

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>