Revert Android Auto support
This PR reverts https://github.com/nextcloud/talk-android/pull/3225

The reason for this is that somehow Google playstore doesn't accept the app anymore when Android Auto support is enabled.
There were multiple emails exchanged with google, but there is no reason given that makes sense why the app is declined.
Some vague hint was that is is related to Android Auto, and indeed removing Android Auto support finally solved that the app was accepted again.

One reason from Google was "We could not access the in-app content with the login credentials that you have provided". But the credentials are provided and working. So it doesn't make sense why/how this is related to Android Auto.

For now i give up to understand what's going wrong with the google playstore policy checks.

Further hint might be at https://developer.android.com/docs/quality-guidelines/car-app-quality.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>