remove FOREGROUND_SERVICE_PHONE_CALL
FOREGROUND_SERVICE_PHONE_CALL is removed because it is not necessary.

Seems it was introduced when moving to Android 14 because NCFirebaseMessagingService originally contained android:foregroundServiceType="phoneCall"

But this foregroundServiceType is not allowed/necessary for NCFirebaseMessagingService so it was removed, but it was forgotten to also remove FOREGROUND_SERVICE_PHONE_CALL permission.

This commit will remove FOREGROUND_SERVICE_PHONE_CALL permission so release process on gplay won't complain.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>