WIP. migrate from logansquare to kotlinx
This commit is breaking the build for now!

There are added some TODOS that need to be resolved if migration is continued!!!!

This migration was initiated because of federation: it seemed to be necessary that RoomLastMessage(json) must be able to be of type ChatMessage and RoomProxyMessage, which is problematic with logansquare. Because of this (and because logansquare is outdated anyway), migration to kotlinx was started.
However it turned out that parsing to RoomProxyMessage is not necessary (for now?!).

So for now, work is paused (and might not be continued as there might be a migration to OpenAPI instead in the future).

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>