manually bump to 19.0.0 Alpha 02
there seems to be something wrong that alpha workflow did not update the apps version.
Thats why gplay complains with
"Google Api Error: forbidden: Cannot update a published APK. - Cannot update a published APK."
because it always tries to upload 19.0.0 Alpha 02.

On nextcloud ftp server, it was most probably always overwritten as there is an up to date alpha2 version.

Seeting the version number to 19.0.0 Alpha 02 should make sure that an alpha 3 is built on next automation run.

The root cause that the version was/is not changed by the automation is still unclear.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>