remove unnecessary KEY_FROM_NOTIFICATION_START_CALL handling
this seems to be a relict caused by conductor back then.

CallNotificationActivity is opened directly.
There shouldn't have been any scenario anymore when it would be opened by MainActivity.
Back then with conductor, this was done because there must have been an activity to open which then opened a controller.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>