manually bump to 20.1.0 Alpha 02
As build script was not able to push to master, it failed and next time it tried to build the same version again.
Thats why Google complained the apk already exists.

The bot was now changed to have admin permissions. So this commit will bump the version number manually and in the next runs everything should work again.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>